requirements.txt
annotated-types==0.7.0
Faker==38.2.0
fastapi>=0.115.0
grpcio>=1.78.0
grpcio-tools>=1.78.0
pydantic==2.12.4
pydantic-core==2.41.5
pyodbc==5.3.0
python-dotenv==1.2.1
sqlalchemy>=2.0.23
strawberry-graphql[asgi]>=0.230.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn>=0.30.0